refactor: рефакторинг stack// refactoring stack
